Moved DualPaneContentState to different package.
