+++ WordPress/src/main/java/org/wordpress/android/ui/DualPaneContentActivity.java
@@ -5,7 +5,6 @@ import android.support.annotation.Nullable;
 import android.support.v4.app.Fragment;
 import android.support.v7.app.AppCompatActivity;
 
-import org.wordpress.android.util.DualPaneContentState;
 import org.wordpress.android.util.DualPaneHelper;
 
 import de.greenrobot.event.EventBus;
